Factor unindent into its own crate
