+++ .cargo/config
@@ -1,0 +1,1 @@
+paths = ["unindent"]
